pyproject.toml
[tool.poetry]
name = "telegram-openai-assistant-bot-exampe"
version = "0.1.0"
description = ""
authors = ["nk0novalov"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
openai = "^1.30.3"
python-telegram-bot = "^21.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

